+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.0] - 2025-09-21
+
+### Added
+
+- Support BiliBili video summary
+
+### Fixed
+
+- Onboarding screen background color
+
 ## [0.7.3] - 2025-09-16
 
 ### Fixed